Fix: move dto to domain
