+++ Source/Domain/Iwentys.Domain/AccountManagement/Mentors/Dto/GroupMentorsDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Iwentys.Modules.AccountManagement.Mentors.Dtos
+namespace Iwentys.Domain.AccountManagement.Mentors.Dto
 {
     public class GroupMentorsDto
     {
