+++ Source/Domain/Iwentys.Domain/AccountManagement/Mentors/Dto/SubjectMentorsDto.cs
@@ -1,11 +1,9 @@
 ﻿using System.Collections.Generic;
-using Iwentys.Modules.AccountManagement.Mentors.Dtos;
 
-namespace Iwentys.Infrastructure.Application.Modules.AccountManagment.Dtos
+namespace Iwentys.Domain.AccountManagement.Mentors.Dto
 {
     public record SubjectMentorsDto
     {
-        // int Id, string Name, IReadOnlyList<GroupMentorsDto> Groups   
         public int Id { get; set; }
         public string Name { get; set; }
         public IReadOnlyList<GroupMentorsDto> Groups { get; set; }
